déboguage dessin Rectangle

+++ src/polyFever/module/affichage/Affichage.java
@@ -11,8 +11,8 @@ public class Affichage {
 	
 	private final int WIDTH; // Largeur de la fenetre en pixel
 	private final int HEIGHT; // Hauteur de la fenetre en pixel
-	private int RATIOPIXWIDTH; // Largeur d'un pixel en float
-	private int RATIOPIXHEIGHT; // Hauteur d'un pixel en float
+	private float RATIOPIXWIDTH; // Largeur d'un pixel en float
+	private float RATIOPIXHEIGHT; // Hauteur d'un pixel en float
 	
 
 
@@ -20,8 +20,9 @@ public class Affichage {
 	{
 		this.WIDTH = width;
 		this.HEIGHT = height;
-		RATIOPIXWIDTH = 2/width;
-		RATIOPIXHEIGHT = 2/height;
+		RATIOPIXWIDTH = 2/(float) width;
+		RATIOPIXHEIGHT = 2/(float) height;	
+		
 		dJeu = new DessinJeu(width, height, this, p, partie);
 		dMenu = new DessinMenu();
 	}
@@ -36,12 +37,12 @@ public class Affichage {
 		return HEIGHT;
 	}
 	
-	public int getRATIOPIXWIDTH() {
+	public float getRATIOPIXWIDTH() {
 		return RATIOPIXWIDTH;
 	}
 
 
-	public int getRATIOPIXHEIGHT() {
+	public float getRATIOPIXHEIGHT() {
 		return RATIOPIXHEIGHT;
 	}
 
@@ -59,5 +60,4 @@ public class Affichage {
 		dMenu.init();
 	}
 	
-
 }
